Add first route to index
*Add first route with some flavor text
*Add shotgun for development
*I use an enviroment variable that is set to nothing in development,
however when it gets to the server it's going to need this available so
that it can be served in a subdomain from the main page. The reason why
for that is because in production I am serving multiple small ruby apps
on the same server that are all accessible from the top level domain.