Add folder names to index
Use pathname to be able to clooect the names of all sub folders in a
given directory and display them on the index view. Use dotenv to set
the path between environments. There's a JobFolderHelper class because
we will need to blacklist and sort out some directories from the group
in the future.