Add SecureShell class and net-ssh gem
Adds a SecureShell class that will use net-ssh to allow ssh access to
servers. This is how our web app will access our server to create the
requested folder. Our server's permissions restict access to the Jobs
folder via smb. Currently SecureShell simply outputs the generated
command we would run on the server. The code that is currently commented
out will be what we use to actually connect and run the generated
command.