Hide file names & protected directories from view
Up until now, files were presented along with diretories on the index
view. We want to hide them. We also have a list of protected directories
we would rather not show either so now we filter those out. Using
environment variables appears to not work unless we actually instantiate
the object, so now we instantiate JobFolderHelpers. We also maintain
seperate methods for view presentation and folder name validation
because while we don't want to show certain items in the view we still
want to protect them from being overwritten.