Feature: Server Restart Button
Added a view accesible only by typing in the address, and with an
authentication layer that has a button to restart the facility file
server. This is so that when the server goes down there is an easier way
to restart it for approved people not in the IT department. Let's see
how this goes! 0_o

Add spacing under buttons

They looked a little cramped.

Feature: Server Restart Button

Added a view accesible only by typing in the address, and with an
authentication layer that has a button to restart the facility file
server. This is so that when the server goes down there is an easier way
to restart it for approved people not in the IT department. Let's see
how this goes! 0_o