Change restart command
Restart command wasn't working. This is probably because I removed 'now'
from the command orignally sent to me by the C.E. Keeping now crashes
the app because ssh terminates suddenly. Removing 'now' would schedule
it for 1 minute from now and allow the app to exit ssh gracefully. While
this worked on my CentOS test server it seems to not work on FreeBSD. So
we will add now back in so the restart button performs as intended and
for now let the app crash until we can find a more graceful way to
redirect in the browser.