Fix connectipon closed by remote host for Net SSH
After an update to the remote FreeBSD server, it now immediatley closes
the connection if you try to connect with a 'none' encryption algorithm.
So now we specify our own list of algorithms so we can omit it. This
also reduces packet sizes. Updated to latest gem to stay ahead of future
updates.